pyproject.toml
[tool.poetry]
name = "gamezone-app"
version = "0.1.0"
description = ""
authors = ["Vachagan Grigoryan <vachagan.grigoryan@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10.10"
aniso8601 = "9.0.1"
asgiref = "3.7.2"
Django = "4.2.2"
promise = "2.3"
Rx = "3.2.0"
singledispatch = "4.0.0"
six = "1.16.0"
sqlparse = "0.4.4"
text-unidecode = "1.3"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.1"
Pillow = "^9.5.0"
django-filter = "^23.2"
Markdown = "^3.4.3"
black = "^23.3.0"
isort = "^5.12.0"
strawberry-graphql = "^0.186.0"
strawberry-graphql-django = "^0.9.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
